properly dispose of previous form state.
This uses an internal MST API to detect we are attaching state to the
same node. I'm not very happy about this so I reported it to MST. I also
wanted to use IDisposer which is described in the MST docs but not
exposed for reuse, so I reported that as a bug too.