Implement dynamic.
Change the infrastructure around the decimal converter so that it
expects all its options. Write code so that we can create a
converter with default options already filled in. This allows us to
remove custom code for decimal option mangling. Add a dynamic converter
which can be used to these partial options dynamically.

It doesn't work with an accessor yet, only context, so we haven't
actually added any new feature yet.