Clean up the way conversion errors are generated.
This in preparation of having multiple types of conversion errors.
We have removed 'rawValidate' and 'validate' from the converter
options, because they weren't pulling their weight - we weren't
using them. Instead we introduce a ConversionError class and we
center the conversion behavior around that. If you throw a
ConversionError the conversion failed.