Don't double-initialize form accessors.
At some point we started initializing the form accessor straight
away in its constructor, but continued to initialize it after
construction. This created a double initialization of an accessor. This
could be compounded if a repeating form accessor was in use. This resulted
in too many derived reactions to be created.