We started moving tracking 'access' states in the accessors themselves.
Repeating forms never worked quite well for this, due to the nature of
paths. By passing them back in backend processing, we can start
to track this in the accessors themselves, which is more resilient.
It also allows the backend to control which fields are accessible
in response to changes in the form.

This doesn't implement it for all accessor types yet.